Authorization in frontend

+++ bk/books.js
@@ -1,3 +1,5 @@
+const jwt = require("jsonwebtoken")
+
 function booksSQLite(app, db) {
   db.run(`
   CREATE TABLE IF NOT EXISTS books (
@@ -8,6 +10,7 @@ function booksSQLite(app, db) {
   `);
 
   app.get('/api/books', (req, res) => {
+    if (!verifyToken(req, res)) return
     const sql = "SELECT * FROM books"
     db.all(sql, (error, result) => {
       if (error) return res.json(error.message)
@@ -16,6 +19,7 @@ function booksSQLite(app, db) {
   })
 
   app.get('/api/books/:id', (req, res) => {
+    if (!verifyToken(req, res)) return
     const { id } = req.params
     const sql = "SELECT * FROM books WHERE id=?"
     db.get(sql, [id], (error, result) => {
@@ -25,6 +29,7 @@ function booksSQLite(app, db) {
   })
 
   app.post('/api/books', (req, res) => {
+    if (!verifyToken(req, res)) return
     const { name, price } = req.body
     const sql = "INSERT INTO books (name,price) VALUES (?,?)"
     db.run(sql, [name, price], (error) => {
@@ -34,6 +39,7 @@ function booksSQLite(app, db) {
   })
 
   app.post('/api/books/:id', (req, res) => {
+    if (!verifyToken(req, res)) return
     const { id } = req.params
     const { name, price } = req.body
     const sql = "UPDATE books set name=?, price=? WHERE id=?"
@@ -44,6 +50,7 @@ function booksSQLite(app, db) {
   })
 
   app.delete('/api/books', (req, res) => {
+    if (!verifyToken(req, res)) return
     const sql = "DELETE FROM books"
     db.run(sql, (error) => {
       if (error) return res.json(error.message)
@@ -52,6 +59,7 @@ function booksSQLite(app, db) {
   })
 
   app.delete('/api/books/:id', (req, res) => {
+    if (!verifyToken(req, res)) return
     const { id } = req.params
     const sql = "DELETE FROM books WHERE id=?"
     db.run(sql, [id], (error) => {
@@ -59,6 +67,20 @@ function booksSQLite(app, db) {
       return res.json(id + " book deleted")
     })
   })
+
+  function verifyToken(req, res) {
+    const token = req?.headers?.authorization?.split(" ")[1];
+    try {
+      return jwt.verify(token, "RANDOM-TOKEN");
+    } catch (error) {
+      if (error instanceof jwt.TokenExpiredError) {
+        res.status(401).json({ error: "Token has expired" });
+      } else {
+        res.status(500).json({ error: "Token verification failed" });
+      }
+      return false
+    }
+  }
 
 
 }
